# auth/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth",
    "summary": "Сервис авторизации онлайн-кинотеатра",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/api/v1/oauth/{provider}/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Возврат от соцсети",
        "description": "Меняет код поставщика на пару токенов. Если вход начинали с access-токеном, аккаунт привязывается к текущему пользователю, и токены не выдаются; привязка не состоится, если к этому времени тот вход закончился — сессию закрыли или сменили пароль. Первый вход через соцсеть заводит учётную запись без пароля: задать его можно в личном кабинете.",
        "operationId": "oauth_callback_auth_api_v1_oauth__provider__callback_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Одноразовый код от поставщика",
              "default": "",
              "title": "Code"
            },
            "description": "Одноразовый код от поставщика"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Значение, с которым начинали вход",
              "default": "",
              "title": "State"
            },
            "description": "Значение, с которым начинали вход"
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Причина отказа, если поставщик его вернул",
              "title": "Error"
            },
            "description": "Причина отказа, если поставщик его вернул"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialLoginSchema"
                }
              }
            }
          },
          "503": {
            "description": "`oauth_provider_unavailable`",
            "content": {
              "application/json": {
                "examples": {
                  "oauth_provider_unavailable": {
                    "summary": "Social provider is unavailable, try again later",
                    "value": {
                      "code": "oauth_provider_unavailable",
                      "detail": "Social provider is unavailable, try again later"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "401": {
            "description": "`oauth_state_invalid`, `oauth_rejected`, `social_link_expired`",
            "content": {
              "application/json": {
                "examples": {
                  "oauth_state_invalid": {
                    "summary": "Login through the social provider has expired, start again",
                    "value": {
                      "code": "oauth_state_invalid",
                      "detail": "Login through the social provider has expired, start again"
                    }
                  },
                  "oauth_rejected": {
                    "summary": "Social provider did not confirm the login, start again",
                    "value": {
                      "code": "oauth_rejected",
                      "detail": "Social provider did not confirm the login, start again"
                    }
                  },
                  "social_link_expired": {
                    "summary": "The session that started linking is no longer valid, log in and start again",
                    "value": {
                      "code": "social_link_expired",
                      "detail": "The session that started linking is no longer valid, log in and start again"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "409": {
            "description": "`social_account_taken`",
            "content": {
              "application/json": {
                "examples": {
                  "social_account_taken": {
                    "summary": "This social account is already linked to another user",
                    "value": {
                      "code": "social_account_taken",
                      "detail": "This social account is already linked to another user"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "404": {
            "description": "`provider_not_found`",
            "content": {
              "application/json": {
                "examples": {
                  "provider_not_found": {
                    "summary": "Unknown social provider or it is not configured",
                    "value": {
                      "code": "provider_not_found",
                      "detail": "Unknown social provider or it is not configured"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorSchema": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код ошибки для программ",
            "examples": [
              "token_expired"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание для человека",
            "examples": [
              "Access token has expired"
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorSchema",
        "description": "Ошибка."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SocialAccountSchema": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Имя поставщика",
            "examples": [
              "yandex"
            ]
          },
          "social_id": {
            "type": "string",
            "title": "Social Id",
            "description": "Идентификатор аккаунта у поставщика",
            "examples": [
              "1234567890"
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "Имя владельца в соцсети",
            "examples": [
              "Нео"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Почта в соцсети",
            "examples": [
              "neo@example.com"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Когда аккаунт привязан"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "social_id",
          "display_name",
          "email",
          "created_at"
        ],
        "title": "SocialAccountSchema",
        "description": "Аккаунт в соцсети, привязанный к учётной записи."
      },
      "SocialLoginSchema": {
        "properties": {
          "tokens": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenPairSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Пара токенов; null — аккаунт привязан к текущему пользователю"
          },
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocialAccountSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Привязанный аккаунт; null — это был вход"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "Учётная запись заведена этим входом"
          }
        },
        "type": "object",
        "required": [
          "tokens",
          "account",
          "created"
        ],
        "title": "SocialLoginSchema",
        "description": "Итог возврата от поставщика: вход или привязка аккаунта."
      },
      "TokenPairSchema": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT для доступа к API, живёт недолго"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "JWT для получения новой пары, одноразовый"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Тип токена для заголовка Authorization",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Через сколько секунд истечёт access-токен",
            "examples": [
              900
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPairSchema",
        "description": "Пара токенов. access-токен передаётся в заголовке `Authorization: Bearer <token>`."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "accessToken": {
        "type": "http",
        "description": "access-токен из ответа на вход или обновление токенов",
        "scheme": "bearer"
      }
    }
  },
...
}